[refactor] Extracting Locations to its own file to remove circular dependency
