+++ src/Locations.ts
@@ -1,0 +1,90 @@
+import { IMap } from "./types";
+
+/**
+ * Display Location Content when visiting somewhere in the World
+ */
+// tslint:disable-next-line:variable-name
+export const Locations: IMap<string> = {
+    // Cities
+    Aevum: 			"Aevum",
+    Chongqing: 		"Chongqing",
+    Ishima: 		"Ishima",
+    NewTokyo: 		"New Tokyo",
+    Sector12: 		"Sector-12",
+    Volhaven: 		"Volhaven",
+
+    // Aevum Locations
+    AevumAeroCorp:              "AeroCorp",
+    AevumBachmanAndAssociates:  "Bachman & Associates",
+    AevumClarkeIncorporated:    "Clarke Incorporated",
+    AevumCrushFitnessGym:       "Crush Fitness Gym",
+    AevumECorp:                 "ECorp",
+    AevumFulcrumTechnologies:   "Fulcrum Technologies",
+    AevumGalacticCybersystems:  "Galactic Cybersystems",
+    AevumNetLinkTechnologies:   "NetLink Technologies",
+    AevumPolice:                "Aevum Police Headquarters",
+    AevumRhoConstruction:       "Rho Construction",
+    AevumSlums:                 "Aevum Slums",
+    AevumSnapFitnessGym:        "Snap Fitness Gym",
+    AevumSummitUniversity:      "Summit University",
+    AevumTravelAgency:          "Aevum Travel Agency",
+    AevumWatchdogSecurity:      "Watchdog Security",
+
+    // Chongqing locations
+    ChongqingKuaiGongInternational: "KuaiGong International",
+    ChongqingSlums:                 "Chongqing Slums",
+    ChongqingSolarisSpaceSystems:   "Solaris Space Systems",
+    ChongqingTravelAgency:          "Chongqing Travel Agency",
+
+    // Sector 12
+    Sector12AlphaEnterprises:   "Alpha Enterprises",
+    Sector12BladeIndustries:    "Blade Industries",
+    Sector12CIA:                "Central Intelligence Agency",
+    Sector12CarmichaelSecurity: "Carmichael Security",
+    Sector12CityHall:           "Sector-12 City Hall",
+    Sector12DeltaOne:           "DeltaOne",
+    Sector12FoodNStuff:         "FoodNStuff",
+    Sector12FourSigma:          "Four Sigma",
+    Sector12IcarusMicrosystems: "Icarus Microsystems",
+    Sector12IronGym:            "Iron Gym",
+    Sector12JoesGuns:           "Joe's Guns",
+    Sector12MegaCorp:           "MegaCorp",
+    Sector12NSA:                "National Security Agency",
+    Sector12PowerhouseGym:      "Powerhouse Gym",
+    Sector12RothmanUniversity:  "Rothman University",
+    Sector12Slums:              "Sector-12 Slums",
+    Sector12TravelAgency:       "Sector-12 Travel Agency",
+    Sector12UniversalEnergy:    "Universal Energy",
+
+    // New Tokyo
+    NewTokyoDefComm:                "DefComm",
+    NewTokyoGlobalPharmaceuticals:  "Global Pharmaceuticals",
+    NewTokyoNoodleBar:              "Noodle Bar",
+    NewTokyoSlums:                  "New Tokyo Slums",
+    NewTokyoTravelAgency:           "New Tokyo Travel Agency",
+    NewTokyoVitaLife:               "VitaLife",
+
+    // Ishima
+    IshimaNovaMedical:          "Nova Medical",
+    IshimaOmegaSoftware:        "Omega Software",
+    IshimaSlums:                "Ishima Slums",
+    IshimaStormTechnologies:    "Storm Technologies",
+    IshimaTravelAgency:         "Ishima Travel Agency",
+
+    // Volhaven
+    VolhavenCompuTek:                   "CompuTek",
+    VolhavenHeliosLabs:                 "Helios Labs",
+    VolhavenLexoCorp:                   "LexoCorp",
+    VolhavenMilleniumFitnessGym:        "Millenium Fitness Gym",
+    VolhavenNWO:                        "NWO",
+    VolhavenOmniTekIncorporated:        "OmniTek Incorporated",
+    VolhavenOmniaCybersystems:          "Omnia Cybersystems",
+    VolhavenSlums:                      "Volhaven Slums",
+    VolhavenSysCoreSecurities:          "SysCore Securities",
+    VolhavenTravelAgency:               "Volhaven Travel Agency",
+    VolhavenZBInstituteOfTechnology:    "ZB Institute of Technology",
+
+    // Generic locations
+    Hospital:               "Hospital",
+    WorldStockExchange:     "World Stock Exchange",
+};
